add a benchmark suite to test performance improvements.
the suite focuses on regex validation for now, since this is a known
hotspot.  the "long" case is similar to regexes required by AWS xray
documents; the "short" case is a modified version of the id-or-name
test schema in typify/tests.

the toolchain version is bumped forward one version so i can compile
on my machine (an Apple M1 Max), but not further because the latest
toolchain regresses several test cases due to stricter irrefutable-if-let
rules.

see https://github.com/oxidecomputer/typify/issues/682